subgraph: refactor & add comments
